refactor: split service mutation methods
